+++ actix-web-lab/src/json.rs
@@ -13,7 +13,6 @@ use actix_web::{
     dev::Payload, error::JsonPayloadError, http::header, web, Error, FromRequest, HttpMessage,
     HttpRequest,
 };
-use derive_more::{Deref, DerefMut, Display};
 use futures_core::Stream as _;
 use serde::de::DeserializeOwned;
 use tracing::debug;
@@ -56,8 +55,33 @@ pub const DEFAULT_JSON_LIMIT: usize = 2_097_152;
 ///     format!("Welcome {}!", info.username)
 /// }
 /// ```
-#[derive(Debug, Deref, DerefMut, Display)]
+#[derive(Debug)]
+// #[derive(Debug, Deref, DerefMut, Display)]
 pub struct Json<T, const LIMIT: usize = DEFAULT_JSON_LIMIT>(pub T);
+
+mod waiting_on_derive_more_to_start_using_syn_2_due_to_proc_macro_panic {
+    use super::*;
+
+    impl<T: std::ops::Deref, const LIMIT: usize> std::ops::Deref for Json<T, LIMIT> {
+        type Target = T;
+
+        fn deref(&self) -> &Self::Target {
+            &self.0
+        }
+    }
+
+    impl<T: std::ops::DerefMut, const LIMIT: usize> std::ops::DerefMut for Json<T, LIMIT> {
+        fn deref_mut(&mut self) -> &mut Self::Target {
+            &mut self.0
+        }
+    }
+
+    impl<T: std::fmt::Display, const LIMIT: usize> std::fmt::Display for Json<T, LIMIT> {
+        fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+            std::fmt::Display::fmt(&self.0, f)
+        }
+    }
+}
 
 impl<T, const LIMIT: usize> Json<T, LIMIT> {
     /// Unwraps into inner `T` value.